+++ Assets/HDD/Scripts/DesktopRenderer.cs
@@ -1,0 +1,17 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+namespace HyperDesktopDuplication {
+  public class DesktopRenderer : MonoBehaviour {
+    public bool visible { get; private set; }
+
+    void OnBecameVisible() {
+      this.visible = true;
+    }
+
+    void OnBecameInvisible() {
+      this.visible = false;
+    }
+  }
+}