+++ Assets/HDD/Scripts/Utils.cs
@@ -1,0 +1,9 @@
+namespace HyperDesktopDuplication {
+  public static class Logger {
+    static string prefix = "HyperDesktopDuplication: ";
+
+    public static void Log(object message) {
+      UnityEngine.Debug.Log(prefix + message);
+    }
+  }
+}