+++ Assets/Scripts/App.cs
@@ -1,11 +1,10 @@
-using System;
 using DT.UniStart;
-using uDesktopDuplication;
+using HyperDesktopDuplication;
 using UnityEngine;
 using UnityEngine.SceneManagement;
 
 public class App : Entry {
-  void Awake() {
+  async void Awake() {
     // show on all desktops
 #if !UNITY_EDITOR
     if (Config.instance.ShowOnAllDesktops) {
@@ -17,30 +16,34 @@ public class App : Entry {
 #endif
 
     // inject context
+    var manager = this.Add(this.GetComponentInChildren<HDD_Manager>());
 #if UNITY_EDITOR
     var eb = this.Add<IEventBus>(new DebugEventBus());
 #else
     var eb = this.Add<IEventBus>(new EventBus());
 #endif
 
+    await manager.Refresh();
+    eb.Invoke("hdd.manager.initialized");
+
     this.onUpdate.AddListener(() => {
       // Ctrl + S to save config
       if (Input.GetKey(KeyCode.LeftControl) && Input.GetKeyDown(KeyCode.S)) {
         // get monitor info
         var old = Config.instance.Monitors;
-        Config.instance.Monitors = new MonitorConfig[Manager.monitors.Count];
-        for (var i = 0; i < Manager.monitors.Count; i++) {
+        Config.instance.Monitors = new MonitorConfig[manager.Monitors.Count];
+        for (var i = 0; i < manager.Monitors.Count; i++) {
           Config.instance.Monitors[i] = new MonitorConfig();
           Config.instance.Monitors[i].Show = false;
         }
-        GameObject.Find("MonitorManager").GetComponentsInChildren<uDesktopDuplication.Texture>().ForEach((texture) => {
-          var config = Config.instance.Monitors[texture.monitorId];
+        GameObject.Find("MonitorManager").GetComponentsInChildren<HDD_Monitor>().ForEach((monitor) => {
+          var config = Config.instance.Monitors[monitor.id];
           config.Show = true;
-          config.Position = texture.transform.position;
-          config.Rotation = texture.transform.rotation.eulerAngles;
-          config.Scale = texture.transform.localScale;
-          config.Bend = texture.bend;
-          config.BendRadius = texture.radius;
+          config.Position = monitor.transform.position;
+          config.Rotation = monitor.transform.rotation.eulerAngles;
+          config.Scale = monitor.transform.localScale;
+          // config.Bend = texture.bend;
+          // config.BendRadius = texture.radius;
         });
 
         Config.Save();